feat(loans): #316 useLoanPortfolioSummary — reemplazar setSelectedLoans del store
Migracion mas grande de Fase 2 del lado creditos. Reemplaza la maquinaria
manual del store (batches de 5, tokens, AbortController, retry, progress)
por TanStack Query con useQueries — N per-loan queries con cache,
dedupe y abort automaticos.

src/queries/loans.ts (nuevo):
- useLoanCashflow(loanId, type, filterDate) — query individual con
  staleTime 5min y retry 0.
- useLoanPortfolioSummary(selectedLoans, filterDate) — hook compuesto
  que fan out via useQueries y agrega via useMemo.
- buildPortfolioSummary movido aqui como helper puro (verbatim del
  store; mismas formulas para weighted IRR/tenor/duration por banco).
- mergeCashflowsByDate idem.
- retryFailed() invalida queryClient solo para los failedLoanIds.

src/pages/loans/index.tsx:
- Reemplaza useAppStore({cashFlows, mergedCashFlows, fullLoan,
  calculationProgress, setSelectedLoans, retryFailedLoans}) con un solo
  useLoanPortfolioSummary(selectedLoanRows, filterDate).
- Elimina el sync useEffect que pusheaba selectedIds al store —
  selectedLoanRows se computa local con useMemo.

src/store/loans/index.ts: -533 lineas netas.
Eliminado:
- cashFlows, mergedCashFlows, fullLoan, loanDebtData, chartData,
  calculationProgress (datos remotos + derivados, ahora en queries).
- selectedLoans field (UI selection vive en pages/loans/index.tsx).
- setSelectedLoans (la action del flicker), setMergedCashFlows,
  getLoanChartData, setCashFlowItem, retryFailedLoans.
- buildPortfolioSummary helper (movido).
- Tokens, AbortController, isStaleCalc, startCalc helpers.

Mantenido (write-side):
- banks, loans (lookup data).
- showDeleteConfirm/Modal/etc (UI flags).
- createLoan, deleteLoanItem, deleteMultipleLoans, getLoanData,
  setSelectedBanks, setFilterDate, setCurrentSelection, wakeServer.

Beneficios:
- Re-seleccionar mismo loan: instant (cache 5min).
- Cancelar al cambiar seleccion: AbortSignal del query lifecycle.
- Failure aislada: 1 loan que falla no afecta los 67 demas (semantica
  identica a #294 pero gratis).
- Devtools: ver todas las queries [loans, cashflow, ...] activas.

Closes #316
Refs #299
